requirements.txt
apturl==0.5.2
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==1.0.0
Django==4.1.1
django-bootstrap4==22.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.8.0
future==0.18.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.2
protobuf==3.6.1
psycopg2==2.9.3
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.16.5
wadllib==1.3.3
whitenoise==6.2.0
xkit==0.0.0
